requirements.txt
Pillow
Django
django-rest-framework
tensorflow>=2.0.0